### 2018-06-19 #784 @leeper
@thomasp85 would you be willing to review this for JOSS?
