### 2018-06-19 #784 @thomasp85
Yes - but I have to be honest about my calendar and say that it will happen earliest in august... If that doesn't work then you'll have to ping someone else 🙂
