### 2018-06-20 #784 @leeper
@nhejazi Are you willing to review this for JOSS?
