### 2018-06-20 #784 @nhejazi
@leeper --- sure thing; I would be happy to review this submission for JOSS.
