### 2018-06-20 #786 @whedon
[REVIEW]: iml: An R package for Interpretable Machine Learning
**Submitting author:** @christophM (<a href="http://orcid.org/0000-0003-2331-868X">Christoph Molnar</a>)
**Repository:** <a href="https://github.com/christophM/iml" target ="_blank">https://github.com/christophM/iml</a>
**Version:** v0.5.2
