### 2018-06-20 #784 @leeper
@nhejazi Thanks! That's great. The review issue is open for you here: https://github.com/openjournals/joss-reviews/issues/786

Please let me know (over there) if you have any questions.
