### 2018-06-21 christophM/iml#23 @nhejazi
[JOSS] package outdated in Binder
In the script `install.R`, the package `iml` is installed from CRAN using `install.packages()`, which installs v0.5.1 of the package; however, the Jupyter notebook available via Binder includes an example with `Interaction`, which is not available in v0.5.1 of the package. This might prove confusing to new users, and may be resolved in one of two ways: (1) updating the CRAN version of `iml` to v0.5.2, or (2) installing the `iml` package from GitHub in `install.R` by includingt he following:

```r
install.packages("remotes")
remotes::install_github("christophM/iml")
```

This is an outstanding item with respect to the [ongoing JOSS review](https://github.com/openjournals/joss-reviews/issues/786) with respect to the items on "Functionality" and "Example Usage". Please feel free to let me know if there are any questions about this.
