### 2018-06-21 christophM/iml#24 @nhejazi
[JOSS] package licensing and contributions
As part of the [ongoing JOSS review](https://github.com/openjournals/joss-reviews/issues/786), this package is required to include a plain-text version of an approved open source license as well as guidelines for contributors.

__License:__
Due to CRAN's policies regarding the MIT license, the file `LICENSE` cannot actually include the text of the MIT license; however, it is still possible to include the text of this license in the package. For example, one might include the text somewhere in the file `README.Rmd` as is done, for example, [here](https://github.com/benkeser/drtmle#license). This would then allow users to refer to the terms under which the `iml` package is distributed without actually looking up the text of the MIT license.

__Contributions:__
The package needs to include clear guidelines for interested contributors. This can take many forms --- a common solution is to include a new file `CONTRIBUTING.md` as is done, for example, [here](https://github.com/tlverse/sl3/blob/master/CONTRIBUTING.md) with an accompanying note in the `README.Rmd`. Of course, many other solutions are acceptable as well.

These are outstanding items with respect to the JOSS review, but can be resolved very quickly. Please let me know if there are any questions about this.
