### 2018-06-21 #786 @nhejazi
I've now completed a preliminary review of the `iml` R package --- @christophM, this looks really great! There are just a few outstanding items that require your attention before I can check off the remaining items required for the JOSS review. Please take a look at the two issues ([on examples](https://github.com/christophM/iml/issues/23) and [on licensing/contributions](https://github.com/christophM/iml/issues/24)) that provide specific details on what needs to be addressed. Feel free to let me know if I can provide any further information that might be helpful in completing this review.
