### 2018-06-22 #786 @leeper
@nhejazi Thank you for the very fast review!

@christophM Please address these issues when you get a chance and let me know when they're complete.
