+++ install.R
@@ -2,3 +2,5 @@ install.packages("mlr")
 install.packages("randomForest")
 install.packages("iml")
 install.packages("gower")
+install.packages("remotes")
+remotes::install_github("christophM/iml")
