### 2018-06-25 5969839
adds MIT license to the readme, #24
