### 2018-06-25 christophM/iml#23 @christophM
Done. I chose option (2). Thanks for that suggestion, because option (2) is quite convenient.
