### 2018-06-25 92c0c9c
removes cran install, #23
