### 2018-06-25 7841e8e
adds contribution.md, #24
