### 2018-06-25 9b676ff
links contribution.md in readme, #24
