### 2018-06-25 christophM/iml#24 @christophM
Done, the repo now contains the MIT license text in the readme and links to the contribution.md
