### 2018-06-25 #786 @christophM
@nhejazi  thank you for the review!

@leeper, @nhejazi I adressed and closed the issues (was not sure if to leave them open or close them. feel free to re-open if something is missing).
