### 2018-06-25 8c08328
adds contribution.md link also in .rmd, #24
