### 2018-06-25 #786 @nhejazi
Everything looks good to me! I've checked off the remaining review items.
