### 2018-06-25 #786 @leeper
@christophM The review process is now complete. To finalize your submission and accept your paper in JOSS, we need two things. First, can you confirm that all references in your bibliography have a DOI (if one exists). It looks like a few of those papers should have DOIs.

Second, we need you to deposit a copy of your software repository (including any revisions made during the JOSS review process) with a data-archiving service. To do so:

 3. Post a comment here with the DOI for the release.
