### 2018-06-27 6035233
adds dois for JOSS paper
