+++ paper/paper.bib
@@ -61,7 +61,8 @@
   number={3},
   pages={916--954},
   year={2008},
-  publisher={Institute of Mathematical Statistics}
+  publisher={Institute of Mathematical Statistics},
+  doi = {10.1214/07-AOAS148}
 }
 
 
@@ -87,7 +88,8 @@ year = {2014}
   booktitle={Proceedings of the 22nd ACM SIGKDD International Conference on Knowledge Discovery and Data Mining},
   pages={1135--1144},
   year={2016},
-  organization={ACM}
+  organization={ACM},
+  doi = {10.1145/2939672.2939778}
 }
 
 
@@ -118,6 +120,6 @@ year = {2018}
   journal={Annals of statistics},
   pages={1189--1232},
   year={2001},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi = {10.1214/aos/1013203451 }
 }
-
