### 2018-06-27 #786 @christophM
I went through the references and updated the DOIs, wherever I found one.
