### 2018-06-27 #786 @christophM
Here is the DOI:

https://doi.org/10.5281/zenodo.1299059
