+++ README.Rmd
@@ -5,6 +5,9 @@ output: github_document
 [![CRAN Status Badge](http://www.r-pkg.org/badges/version/iml)](https://CRAN.R-project.org/package=iml)
 [![CRAN Downloads](http://cranlogs.r-pkg.org/badges/grand-total/iml)](https://cran.rstudio.com/web/packages/iml/index.html)
 [![codecov.io](https://codecov.io/github/christophM/iml/coverage.svg?branch=master)](https://codecov.io/github/christophM/iml?branch=master)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1299059.svg)](https://doi.org/10.5281/zenodo.1299059)
+
+
 
 # iml: interpretable machine learning
  `iml ` is an R package that interprets the behaviour and explains predictions of machine learning models.
