### 2018-07-02 #786 @christophM
Hi @arfon, I just talked to @leeper about the possibilty to add an author to the paper after publishing.
When I wrote the paper, I (falsely) assumed that only code contributors can be authors.

It would be really great if I could update the paper with authors that also contributed conceptually.

@leeper mentioned that it might be possible.
@arfon is it possible to update the list of authors?

Thanks for your efforts!
