### 2018-07-03 #786 @arfon
> @arfon is it possible to update the list of authors?

Yes, please update the `paper.md` in your `master` branch of the repository and give me a shout when you've done so.
