### 2018-07-03 #786 @christophM
Thanks @arfon.
I updated the `paper.md` in the `master`. Let me know if there are any more actions I have to take.
