### 2018-07-03 #786 @arfon
> I updated the paper.md in the master. Let me know if there are any more actions I have to take.

👍 this should be updated on the JOSS site now. It can take a few hours to update due to caching.
